requirements.txt
# UI
streamlit>=1.30

# Deep Learning
torch
transformers

# Semantic similarity
sentence-transformers
scikit-learn
numpy

# Image processing
pillow

# Evaluation metrics
nltk
rouge-score

# LLM reasoning (Gemini)
google-generativeai
